Removed special boundary patch difference function.
This function is not necessary. Since we compute the offsets as valid pixels
(unmasked and inside the image) of the target patch, this set of pixels
is always valid in the source patches, since the source patches are always
full and entirely valid. Therefore the logic of which pixels to use is
already in the ComputeOffsets function, so we dont need to do anything
special in the PatchDifference function.

Also use initialization list instead of manual initializations in
constructor.