Required image and mask in PatchBasedInpainting constructor
In order to do this, the PatchBasedInpainting object must be a pointer member
in PatchBasedInpaintingGUI as it cannot be constructed until the images to use
are determined.

Removed ClusterColors dependency from PatchBasedInpainting. Instead, favor the
idea of passing only named images to use in the patch comparison function.
That is, provide a MembershipImage which already contains the result of any
type of clustering.