Remove UpdateMask function
This function was already not being used. It is not needed because now we
copy the patch into the target region generally in all images that are set to
track the inpainting. The source patch is always entirely valid, so the region
that overlaps the target region will always be filled with valid pixels in the
mask.