Major refactoring. NOTE: this commit is for backup purposes only - the code is in a very broken state.
Trying to make code more object oriented.

PairDifferences class introduced to hold all of the differences between one source/target pair.

CandidatePairs class introduced to handle determining which patches are valid source patches.

Store a vector of smartpointers of PatchPairs because this is the only way to store a vector
of an object with const members.