Derive FeatureVector and ImagePatch visitors from InpaintingVisitorParent.
This was done so that they can be stored in a container in a composite visitor as
InpaintingVisitorParent pointers.

To do this, the functions could not be function templates (because
function templates cannot be virtual) . Instead,
the graph type was added to the list of class template paramaters.
The vertex descriptor type is derived from the graph type.