Restructure so that a different image can be used to the histogram test and the SSD comparison.
The PatchInpainters now do the inpainting. The CompositePatchInpainter
allows us to fill multiple images simultaneously - so we can inpaint
say an RGB and an HSV image at the same time. If those images are being
used by other parts of the code, they both need to be filled to stay
current.